Refactor: pathname 처리 로직 수정
